requirements.txt
tabulate
datetime
pandas
mysql-connector-python>=8.0